requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
googlesearch-python>=1.0.0
colorama>=0.4.4